+++ tests/V1/Requirements/RequireAllOfTest.php
@@ -77,7 +77,7 @@ class RequireAllOfTest extends PHPUnit_Framework_TestCase
     /**
      * @covers ::__construct
      */
-    public function testIsRequirement()
+    public function test_is_Requirement()
     {
         // ----------------------------------------------------------------
         // setup your test
@@ -165,7 +165,7 @@ class RequireAllOfTest extends PHPUnit_Framework_TestCase
     /**
      * @covers ::__construct
      * @covers ::apply
-     * @expectedException GanbaroDigital\Defensive\V1\Exceptions\BadRequirements
+     * @expectedException GanbaroDigital\Defensive\V1\Exceptions\EmptyRequirementsList
      */
     public function testRequirementsArrayCannotBeEmpty()
     {