+++ tests/V1/Requirements/RequireAnyOneOfTest.php
@@ -77,7 +77,7 @@ class RequireAnyOneOfTest extends PHPUnit_Framework_TestCase
     /**
      * @covers ::__construct
      */
-    public function testIsRequirement()
+    public function test_is_Requirement()
     {
         // ----------------------------------------------------------------
         // setup your test
@@ -164,7 +164,7 @@ class RequireAnyOneOfTest extends PHPUnit_Framework_TestCase
 
     /**
      * @covers ::__construct
-     * @expectedException GanbaroDigital\Defensive\V1\Exceptions\BadRequirements
+     * @expectedException GanbaroDigital\Defensive\V1\Exceptions\EmptyRequirementsList
      */
     public function testArrayOfRequirementsCannotBeEmpty()
     {